use factories for preview/editor components
